+++ CHANGELOG.md
@@ -26,6 +26,7 @@
  - New article by [Matthew Larkin](https://github.com/matthewlarkin) about [migrations](https://sql.ophir.dev/your-first-sql-website/migrations.sql).
  - Add a row-level `id` attribute to the button component.
  - Static assets (js, css, svg) needed to build SQLPage are now cached individually, and can be downloaded separately from the build process. This makes it easier to build SQLPage without internet access. If you use pre-built SQLPage binaries, this change does not affect you.
+ - New `icon_after` row-level property in the button component to display an icon on the right of a button (after the text).
 
 ## 0.21.0 (2024-05-19)